Move logging interceptors to `:testutil-logging`
