+++ testutil-logging/src/main/kotlin/io/spine/logging/testing/Recorder.kt
@@ -24,7 +24,7 @@
  * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
 
-package io.spine.testing.logging
+package io.spine.logging.testing
 
 import io.spine.logging.Level
